GCS:UAVOWidgetUtils: Add widget helper functions
Before disabling widgets with UAVO relations from outside ConfigTaskWidget was problematic because ConfigTaskWidget would ignore what you had done and vice-versa. ConfigTaskWidget::setWidgetEnabled to the rescue.

Setting dynamic properties on widgets that are used in the widget stylesheet is problematic because they stylesheet isn't re-evaluated automatically. This provides a small helper function to handle this rather than re-writing it dozens of times elsewhere.